feat(playground): exposing ability to upload images
