+++ packages/playground/.env
@@ -24,3 +24,4 @@ VITE_ENABLE_MODEL_SELECT=true
 VITE_ENABLE_TOOLS=false
 VITE_ENABLE_KNOWLEDGE=false
 VITE_ENABLE_NEW_KNOWLEDGE=true
+VITE_ENABLE_ATTACHMENT=true
