+++ packages/playground/src/components/prompt/index.ts
@@ -1,1 +1,1 @@
-export * from './PromptLibraryComponent';
+export * from "./PromptLibrary";
