+++ packages/playground/src/components/room/index.ts
@@ -1,4 +1,4 @@
-export * from './RoomControlsComponent';
-export * from './RoomAppComponent';
-export * from './RoomInputComponent';
-export * from './RoomMessageComponent';
+export * from "./RoomApp";
+export * from "./RoomControls";
+export * from "./RoomInput";
+export * from "./RoomMessage";
