+++ packages/playground/src/types.d.ts
@@ -95,7 +95,7 @@ interface InputTextPixelMessage extends AbstractPixelMessage {
 	type: "INPUT_TEXT";
 	visible: true;
 	inputUIPrompt: string;
-    modelId: string;
+	modelId: string;
 	paramMap: {
 		max_new_tokens: number;
 		temperature: number;
@@ -121,7 +121,7 @@ interface InputToolExecPixelMessage extends AbstractPixelMessage {
 			/** Parameters for app */
 			map: Record<string, unknown>;
 		};
-	};
+	}[];
 }
 
 interface ResponseTextPixelMessage extends AbstractPixelMessage {
