+++ packages/playground/vite.config.ts
@@ -1,6 +1,6 @@
-import { resolve } from "node:path";
 import react from "@vitejs/plugin-react";
 import { defineConfig, loadEnv } from "vite";
+import { resolve } from "node:path";
 
 export default defineConfig(({ mode }) => {
 	const env = loadEnv(mode, process.cwd(), "");
